changed deletFolderContent to not delete .stickyfolder file

+++ common.py
@@ -67,7 +67,7 @@ def deleteFolderContent(folder):
     for the_file in os.listdir(folder):
         file_path = os.path.join(folder, the_file)
         try:
-            if os.path.isfile(file_path):
+            if os.path.isfile(file_path) and the_file != ".stickyfolder":
                 os.unlink(file_path)
             elif os.path.isdir(file_path): shutil.rmtree(file_path)
         except Exception as e:
